Stop subclassing deprecated Alignment in MultipleSeqAlignment.
Removes the original Alignment object which was defined in the
deprecated Bio.Align.Generic module, so the base methods defined
there need to be added to MultipleSeqAlignment.

Note the docstring examples were updated to use the newer class
name MultipleSeqAlignment.
