+++ Tests/run_tests.py
@@ -77,7 +77,6 @@ VERBOSITY = 0
 # those modules here. Please sort names alphabetically.
 DOCTEST_MODULES = [
     "Bio.Align",
-    "Bio.Align.Generic",
     "Bio.Align.Applications._Clustalw",
     "Bio.Align.Applications._ClustalOmega",
     "Bio.Align.Applications._Dialign",
